requirements.txt
torchgeometry>=0.1.2
PyYAML==5.1.1
configargparse
smplx
pillow
opencv-python
tqdm